+++ libs/core/langchain_core/memory/__init__.py
@@ -1,0 +1,13 @@
+from langchain_core.memory.base import BaseMemory
+from langchain_core.memory.chat_memory import BaseChatMemory
+from langchain_core.memory.chat_message_history import ChatMessageHistory
+from langchain_core.memory.utils import get_prompt_input_key
+
+__all__ = [
+    "BaseChatMemory",
+    "BaseMemory",
+    "ChatMessageHistory",
+    "get_prompt_input_key",
+]
+
+
